+++ R/criterion.R
@@ -28,7 +28,33 @@ root_criterion <- function(testfun, desc) {
 #' @rdname root_criterion
 #' @param x An object
 #' @export
-is.root_criterion <- function(x) inherits(x, "root_criterion")
+is.root_criterion <- function(x) {
+  inherits(x, "root_criterion")
+}
+
+#' @rdname root_criterion
+#' @export
+as.root_criterion <- function(x) UseMethod("as.root_criterion", x)
+
+#' @details
+#' The \code{as.root_criterion} function accepts objects of class
+#' \code{root_criterion}, and character values; the latter will be
+#' converted to criteria using \code{has_file}.
+#'
+#' @rdname root_criterion
+#' @export
+as.root_criterion.character <- function(x) {
+  has_file(x)
+}
+
+#' @rdname root_criterion
+#' @export
+as.root_criterion.root_criterion <- identity
+
+#' @export
+as.root_criterion.default <- function(x) {
+  stop("Cannot coerce ", x, " to type root_criterion.")
+}
 
 #' @export
 format.root_criterion <- function(x, ...) {