+++ R/file.R
@@ -14,8 +14,8 @@
 #' \code{make_find_root_file} which creates a function that will look for the
 #' root every time it is called).
 #'
-#' @param criterion A criterion created by \code{\link{root_criterion}}, e.g.,
-#'   \code{has_file(...)}
+#' @param criterion A criterion, will be coerced using
+#'   \code{\link{as.root_criterion}}
 #' @param path The start directory
 #' @param ... Additional arguments passed to \code{\link{file.path}}
 #' @return The normalized path of the root as specified by the search criteria.
@@ -31,7 +31,7 @@
 #' @seealso \code{\link{find_root}} \code{\link[utils]{glob2rx}} \code{\link[base]{file.path}}
 #'
 #' @export
-find_root_file <- function(..., criterion, path = getwd()) {
+find_root_file <- function(..., criterion, path = ".") {
   root <- find_root(criterion = criterion, path = path)
   file.path(root, ...)
 }