+++ NEWS.md
@@ -1,3 +1,7 @@
+- New function `as.root_criterion` (called by `find_root()`), a character argument is converted to a `has_file()` criterion. Now the call `find_root("filename")` is identical to `find_root(has_file("filename"))`.
+- Use `"."` instead of `"getwd()"` as default for `path` argument
+- Improved documentation
+
 Version 0.0-4 (2015-05-20)
 ===